requirements.txt
Django==1.6
psycopg2==2.5.1
twitter==1.10.2
pytz==2013.8
django-tastypie==0.10.0
django-pipeline==1.3.17
python-mimeparse==0.1.4